chore: update Java version to 11, Android plugin to 8.2.1, and webview_flutter to ^4.0.10 (#32)(#33)
Initially, the application was compiled for Android. To do this, the following were required:

JavaVersion.VERSION_11
id "com.android.application" version "8.2.1" apply false
The dependency webview_flutter: ^4.0.10

It was also necessary to update the about_file_screen.dart file, which uses this library, and to take the library update into account to ensure compatibility with newer versions.